+++ docs/changelog.md
@@ -7,6 +7,7 @@ Numbers in brackets show the issue number in https://github.com/wp-document-revi
 * NEW: Integrate with PublishPress Statuses plugin for custom statuses. (#335)
 * NEW: Accessibility rules states that links to PDF documents should have visible references. Blocks have an explicit switch. (#322)
 * NEW: User pulldowns will show only relevant users. (#321)
+* NEW: Filter 'document_post_thumbnail' used to define the post-thumbnail image size (if not set by theme). (#339)
 * NEW: Filter 'document_use_wp_filesystem' used to serve document (instead of PHP readfile). Irrelevant if the file is compressed on output. (#320)
 * NEW: Filter 'document_internal_filename' for updating internal file name additionally passed the original name. (#319)
 * NEW: Filter 'document_validate_md5' to switch off attachment MD5 format validation. (#318)